refactor(scrape): move GotSportEventScraper body into gotsport.py; shim
Shell 01 Step 3 - Commit B. Pure mechanical move with zero behavior change:

src/scrapers/gotsport.py (+2254 lines, final 2898 lines):
  - Absorbs the entire GotSportEventScraper class body from gotsport_event.py.
  - Module-level CAPTCHA helpers (_CAPTCHA_URL_MARKER, _CAPTCHA_BODY_MARKER,
    _CAPTCHA_SITEKEY_RE, _RECAPTCHA_SITEKEY_FALLBACK_RE, _find_sitekey,
    _extract_captcha_signals) move in alongside EventCaptchaGatedError and
    the EventTeam dataclass.
  - The Commit-A adapter class is collapsed: GotsportScraper now inherits
    directly from ProviderScraper (not from GotSportEventScraper anymore).
    MRO goes from [GotsportScraper, GotSportEventScraper, ProviderScraper,
    ABC, object] to [GotsportScraper, ProviderScraper, ABC, object].
  - __init__ merges the two Commit-A init paths: session/ZenRows/CAPTCHA
    wiring (formerly in GotSportEventScraper.__init__) plus the
    providers-table assertion (formerly in the adapter). Now the "Initialized
    GotsportScraper ..." log line uses the new class name.
  - End-of-module late-imports from gotsport_event.py (added in Commit A to
    break a circular race) are gone — gotsport_event.py is a shim now, so
    gotsport.py no longer needs to reach into it.
  - All three ABC methods land on the single canonical class:
    fetch_event_metadata (real), fetch_teams_by_cohort + resolve_canonical_team_id
    (NotImplementedError stubs with Shell 01 Step 4+6 pointers).

src/scrapers/gotsport_event.py (-2216 lines, final 20-line shim):
  - Back-compat shim: re-exports GotsportScraper as GotSportEventScraper,
    plus EventCaptchaGatedError and EventTeam. The 8 external importer
    scripts continue to work with zero changes.
  - The former lazy in-function ``from src.scrapers.gotsport import
    GotSportScraper`` inside __init__ is gone with the rest of the class
    body; gotsport.py's own __init__ handles session composition internally
    now.

Verification (all green; identical outputs to Commit A):
  - 29 foundation unit tests pass.
  - All 8 GotSportEventScraper importers load cleanly (via the shim).
  - Adapter suite: GotsportScraper instantiates, passes
    isinstance(ProviderScraper), get_provider_scraper returns it for
    /org_event/events/<id>, UnsupportedProviderError on non-matching URLs
    and on unseeded providers rows, stubs raise NotImplementedError.
  - Shim re-exports are object-identical: GotSportEventScraper is
    GotsportScraper, EventCaptchaGatedError/EventTeam same objects.
  - scrape_specific_event.py 45224 -> CAPTCHA detected in ~5s, sitekey
    6Lf7TGog... captured, artifact written, exit 0, not marked scraped.
  - scrape_specific_event.py 42434 --force -> routed through ZenRows, 104+
    team IDs resolved.
  - "Initialized GotsportScraper ..." log confirms the new class name at
    init time.

Net LoC change: -35 (duplicated imports collapsed; adapter plumbing
deleted).

Refs: .turbo/plans/matchbalance-backtest-intake-01-scraper-fix-and-abstraction.md
      Step 3 (Extract ProviderScraper abstraction), completed in two commits.
